+++ packages/textinput/src/index.ts
@@ -1,2 +1,5 @@
-export { default, TextInputProps } from './react/index'
+import './css/index.css'
+import TextInput from './react/index'
+export default TextInput
+export { TextInputProps } from './react/index'
 export * from './vars/index'
