+++ packages/textinput/src/react/index.tsx
@@ -11,7 +11,6 @@ import {
 } from '@pluralsight/ps-design-system-util'
 import React from 'react'
 
-import '../css/index.css'
 import * as vars from '../vars/index'
 
 const styles = {
